backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
SQLAlchemy>=2.0.0
psycopg[binary]>=3.1.0
bcrypt>=4.0.0
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
pandas>=2.0.0
openpyxl>=3.1.0
fpdf2>=2.7.0
